+++ src/js/create-cards.js
@@ -1,27 +1,3 @@
-console.log(`Hi, cards!`);
-import './cocktail-api' ;
-
-import {fetchRandomCocktails} from './cocktail-api';
-
-const cardList = document.querySelector(`.cocktails-list`)
-
-//console.log(cardList);
-
-let quantity;
-
-quantity = 9;
-
-cardList.innerHTML = ``;
-
-//console.log(fetchRandomCocktails(quantity));
-
-let allCards = await fetchRandomCocktails(quantity);
-
-//console.log(allCards);
-
-cardList.insertAdjacentHTML(`beforeend`, createMarkup(allCards));
-
-// createMarkup(allCards);
 
 function createMarkup(arr) {
     return arr.map(({description, drink, drinkThumb, _id})=>
@@ -33,7 +9,7 @@ function createMarkup(arr) {
             <div class="btns-info">
                 <button type="button" class="js-learn-more learn-more-button">Learn more</button>
                 <button type="button" class="js-add-to add-to-button">
-                    <svg class="icon" width="20px" height="16px">
+                    <svg class="icon-heart" width="20px" height="16px">
                         <use href="/img/svg/sprite.svg#icon-heart"></use>
                     </svg>
                 </button>
@@ -43,5 +19,5 @@ function createMarkup(arr) {
     ).join(``);
 };
 
-//export {createMarkup};
+export {createMarkup};
