+++ src/main.js
@@ -8,4 +8,5 @@ import './js/modal-ingredient';
 import './js/search-by-name';
 import './js/pagination';
 // додаткові підключення робіть під цією стрічкою
+import './js/main-start';
 import './js/create-cards';
