+++ src/partials/main-section.html
@@ -3,98 +3,9 @@
 
     <div class="container container-cocktails">
     <h2 class="cocktails-header">Cocktail</h2>
-    <div class="cocktails-list">
-<!--     
-        <div class="cocktail-card" id="">
-            <img class="cocktail-image" src="" alt="" loading="lazy" width="307" height="257"/>
-            <h3 class="cocktail-name">Cocktail name</h3>
-            <p class="cocktail-info"> An iconic Italian cocktail, the Negroni features a balanced blend of gin, Campari, and sweet vermouth, known for its bittersweet flavor and ruby-red color. Served over ice with an orange slice, it offers a robust and refreshing taste experience.</p>
-            <div class="btns-info">
-                <button type="button" class="js-learn-more learn-more-button">Learn more</button>
-                <button type="button" class="js-add-to add-to-button">
-                    <svg class="icon" width="20px" height="16px">
-                        <use href="/img/svg/sprite.svg#icon-heart"></use>
-                    </svg>
-                </button>
-            </div>
+    <h2 class="cocktails-header visually-hidden">Searching results</h2>
+        <div class="cocktails-list">
+    
         </div>
-        
-         <div class="cocktail-card" id="">
-            <img class="cocktail-image" src="" alt="" loading="lazy" width="307" height="257"/>
-            <h3 class="cocktail-name">Cocktail name</h3>
-            <p class="cocktail-info"> An iconic Italian cocktail, the Negroni features a balanced blend of gin, Campari, and sweet vermouth, known for its bittersweet flavor and ruby-red color. Served over ice with an orange slice, it offers a robust and refreshing taste experience.</p>
-            <div class="btns-info">
-                <button type="button" class="js-learn-more learn-more-button">Learn more</button>
-                <button type="button" class="js-add-to add-to-button">
-                    <svg class="icon" width="20px" height="16px">
-                        <use href="/img/svg/sprite.svg#icon-heart"></use>
-                    </svg>
-                </button>
-            </div>
-        </div>
-
-         <div class="cocktail-card" id="">
-            <img class="cocktail-image" src="" alt="" loading="lazy" width="307" height="257"/>
-            <h3 class="cocktail-name">Cocktail name</h3>
-            <p class="cocktail-info"> An iconic Italian cocktail, the Negroni features a balanced blend of gin, Campari, and sweet vermouth, known for its bittersweet flavor and ruby-red color. Served over ice with an orange slice, it offers a robust and refreshing taste experience.</p>
-            <div class="btns-info">
-                <button type="button" class="js-learn-more learn-more-button">Learn more</button>
-                <button type="button" class="js-add-to add-to-button">
-                    <svg class="icon" width="20px" height="16px">
-                        <use href="/img/svg/sprite.svg#icon-heart"></use>
-                    </svg>
-                </button>
-            </div>
-        </div>
-
-         <div class="cocktail-card" id="">
-            <img class="cocktail-image" src="" alt="" loading="lazy" width="307" height="257"/>
-            <h3 class="cocktail-name">Cocktail name</h3>
-            <p class="cocktail-info"> An iconic Italian cocktail, the Negroni features a balanced blend of gin, Campari, and sweet vermouth, known for its bittersweet flavor and ruby-red color. Served over ice with an orange slice, it offers a robust and refreshing taste experience.</p>
-            <div class="btns-info">
-                <button type="button" class="js-learn-more learn-more-button">Learn more</button>
-                <button type="button" class="js-add-to add-to-button">
-                    <svg class="icon" width="20px" height="16px">
-                        <use href="/img/svg/sprite.svg#icon-heart"></use>
-                    </svg>
-                </button>
-            </div>
-        </div>
-
-         <div class="cocktail-card" id="">
-            <img class="cocktail-image" src="" alt="" loading="lazy" width="307" height="257"/>
-            <h3 class="cocktail-name">Cocktail name</h3>
-            <p class="cocktail-info"> An iconic Italian cocktail, the Negroni features a balanced blend of gin, Campari, and sweet vermouth, known for its bittersweet flavor and ruby-red color. Served over ice with an orange slice, it offers a robust and refreshing taste experience.</p>
-            <div class="btns-info">
-                <button type="button" class="js-learn-more learn-more-button">Learn more</button>
-                <button type="button" class="js-add-to add-to-button">
-                    <svg class="icon" width="20px" height="16px">
-                        <use href="/img/svg/sprite.svg#icon-heart"></use>
-                    </svg>
-                </button>
-            </div>
-        </div>
-
-         <div class="cocktail-card" id="">
-            <img class="cocktail-image" src="" alt="" loading="lazy" width="307" height="257"/>
-            <h3 class="cocktail-name">Cocktail name</h3>
-            <p class="cocktail-info"> An iconic Italian cocktail, the Negroni features a balanced blend of gin, Campari, and sweet vermouth, known for its bittersweet flavor and ruby-red color. Served over ice with an orange slice, it offers a robust and refreshing taste experience.</p>
-            <div class="btns-info">
-                <button type="button" class="js-learn-more learn-more-button">Learn more</button>
-                <button type="button" class="js-add-to add-to-button">
-                    <svg class="icon" width="20px" height="16px">
-                        <use href="/img/svg/sprite.svg#icon-heart"></use>
-                    </svg>
-                </button>
-            </div>
-        </div> -->
-        
-
-
-
-
-
     </div>
-    </div>
-
 </section>
